+++ bootique/src/test/java/io/bootique/jopt/JoptCliProviderCommandNameTest.java
@@ -1,5 +1,6 @@
 package io.bootique.jopt;
 
+import io.bootique.application.ApplicationMetadata;
 import io.bootique.application.CommandMetadata;
 import io.bootique.application.OptionMetadata;
 import io.bootique.cli.Cli;
@@ -12,10 +13,8 @@ import org.junit.Test;
 
 import java.util.Arrays;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
@@ -26,13 +25,11 @@ public class JoptCliProviderCommandNameTest {
 
     private BootLogger mockBootLogger;
     private Map<String, Command> commands;
-    private Set<OptionMetadata> options;
 
     @Before
     public void before() {
         this.mockBootLogger = mock(BootLogger.class);
         this.commands = new HashMap<>();
-        this.options = new HashSet<>();
     }
 
     @Test
@@ -86,7 +83,11 @@ public class JoptCliProviderCommandNameTest {
         Optional<Command> mockDefaultCommand = Optional.of(mock(Command.class));
         Optional<Command> mockHelpCommand = Optional.of(mock(Command.class));
         CommandManager commandManager = new DefaultCommandManager(commands, mockDefaultCommand, mockHelpCommand);
-        return new JoptCliProvider(mockBootLogger, commandManager, options, argsArray).get();
+
+        ApplicationMetadata.Builder appBuilder = ApplicationMetadata.builder();
+        commands.values().forEach(c -> appBuilder.addCommand(c.getMetadata()));
+
+        return new JoptCliProvider(mockBootLogger, commandManager, appBuilder.build(), argsArray).get();
     }
 
 }